requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
gspread==6.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
